end vid 03
Did some work on the platform. Also, the video is working in the Y axis but with this new version of Unreal (4.25.3) it works in the X axis.

Lastly, to get the platform to act like it did in the video when turning on Simulate Physics, I had to Lock X rotation (not Y) and change Linear Damping to 1.0 to prevent the platform from falling over when it hit the ground.